feat: give control over the registration of tasks on the CORBA naming service
Syskit itself does not need it anymore, and in tests it's arguably
generating a lot of noise (registering/deregistering tasks on the
name service).

The main driver is to allow parallel execution of unit tests, which
works right now, but generates so many "name service rebinding" messages
as to be useless.